requirements.txt
fastapi
uvicorn
xrpl-py
google-genai
sqlalchemy
python-multipart
pydantic
python-dotenv
